use int instead of long

+++ test-suite/noarbsabr.cpp
@@ -31,7 +31,7 @@ using namespace boost::unit_test_framework;
 
 namespace {
 void checkD0(const Real sigmaI, const Real beta, const Real rho, const Real nu,
-             const Real tau, const unsigned long absorptions) {
+             const Real tau, const unsigned int absorptions) {
 
     Real forward = 0.03; // does not matter in the end
     Real alpha = sigmaI / std::pow(forward, beta - 1.0);
